+++ packages/cli/src/commands/hydrogen/env/list.test.ts
@@ -1,17 +1,13 @@
 import {describe, it, expect, vi, beforeEach, afterEach} from 'vitest';
 import {mockAndCaptureOutput} from '@shopify/cli-kit/node/testing/output';
 import {inTemporaryDirectory} from '@shopify/cli-kit/node/fs';
-import {renderConfirmationPrompt} from '@shopify/cli-kit/node/ui';
 
 import {getStorefrontEnvironments} from '../../../lib/graphql/admin/list-environments.js';
 import {dummyListEnvironments} from '../../../lib/graphql/admin/test-helper.js';
 import {type AdminSession, login} from '../../../lib/auth.js';
-import {
-  renderMissingLink,
-  renderMissingStorefront,
-} from '../../../lib/render-errors.js';
-import {linkStorefront} from '../link.js';
+import {renderMissingStorefront} from '../../../lib/render-errors.js';
 import {runEnvList} from './list.js';
+import {verifyLinkedStorefront} from '../../../lib/verify-linked-storefront.js';
 
 const SHOP = 'my-shop';
 
@@ -29,6 +25,7 @@ vi.mock('../../../lib/auth.js');
 vi.mock('../../../lib/shopify-config.js');
 vi.mock('../../../lib/render-errors.js');
 vi.mock('../../../lib/graphql/admin/list-environments.js');
+vi.mock('../../../lib/verify-linked-storefront.js');
 vi.mock('../../../lib/shell.js', () => ({getCliCommand: () => 'h2'}));
 
 describe('listEnvironments', () => {
@@ -51,6 +48,12 @@ describe('listEnvironments', () => {
     vi.mocked(login).mockResolvedValue({
       session: ADMIN_SESSION,
       config: SHOPIFY_CONFIG,
+    });
+
+    vi.mocked(verifyLinkedStorefront).mockResolvedValue({
+      id: SHOPIFY_CONFIG.storefront.id,
+      title: SHOPIFY_CONFIG.storefront.title,
+      productionUrl: 'https://my-shop.myshopify.com',
     });
 
     vi.mocked(getStorefrontEnvironments).mockResolvedValue(
@@ -98,39 +101,16 @@ describe('listEnvironments', () => {
 
   describe('when there is no linked storefront', () => {
     beforeEach(() => {
-      vi.mocked(login).mockResolvedValue({
-        session: ADMIN_SESSION,
-        config: {
-          ...SHOPIFY_CONFIG,
-          storefront: undefined,
-        },
-      });
+      vi.mocked(verifyLinkedStorefront).mockResolvedValue(undefined);
     });
 
-    it('calls renderMissingLink', async () => {
+    it("doesn't list any environments", async () => {
       await inTemporaryDirectory(async (tmpDir) => {
+        const output = mockAndCaptureOutput();
+
         await runEnvList({path: tmpDir});
 
-        expect(renderMissingLink).toHaveBeenCalledOnce();
-      });
-    });
-
-    it('prompts the user to create a link', async () => {
-      vi.mocked(renderConfirmationPrompt).mockResolvedValue(true);
-
-      await inTemporaryDirectory(async (tmpDir) => {
-        await runEnvList({path: tmpDir});
-
-        expect(renderConfirmationPrompt).toHaveBeenCalledWith({
-          message: expect.arrayContaining([{command: 'h2 link'}]),
-        });
-
-        expect(linkStorefront).toHaveBeenCalledWith(
-          tmpDir,
-          ADMIN_SESSION,
-          {...SHOPIFY_CONFIG, storefront: undefined},
-          expect.anything(),
-        );
+        expect(output.info()).toBe('');
       });
     });
   });