+++ packages/cli/src/lib/transpile/morph/typedefs.ts
@@ -56,15 +56,11 @@ export function generateTypeDefs(sourceFile: SourceFile, code: string) {
 
   const knownGenerics: Record<string, string | undefined> = {
     MetaFunction: 'T',
+    SerializeFrom: 'T',
   };
 
   typedefsFromImports.forEach((typeElements, moduleSpecifier) => {
     for (const typeElement of typeElements) {
-      // We only use this in root.tsx and it's better to
-      // reuse the existing LoaderReturnData, so skip it.
-      if (typeElement === 'SerializeFrom') continue;
-
-      // Note: SerializeFrom also needs generic if we stop skipping it.
       const hasGeneric = !!knownGenerics[typeElement];
 
       typedefs.push(